Move EmsCommon under the Mixins namespace and update the includes
