+++ app/controllers/ems_block_storage_controller.rb
@@ -1,7 +1,7 @@
 class EmsBlockStorageController < ApplicationController
   include Mixins::GenericListMixin
   include Mixins::GenericShowMixin
-  include EmsCommon
+  include Mixins::EmsCommon
   include Mixins::EmsCommonAngular
   include Mixins::GenericSessionMixin
