+++ app/controllers/ems_container_controller.rb
@@ -1,7 +1,7 @@
 class EmsContainerController < ApplicationController
   include Mixins::GenericListMixin
   include Mixins::GenericShowMixin
-  include EmsCommon # common methods for EmsInfra/Cloud/Container controllers
+  include Mixins::EmsCommon # common methods for EmsInfra/Cloud/Container controllers
   include Mixins::EmsCommonAngular
   include Mixins::GenericSessionMixin
   include Mixins::DashboardViewMixin
