+++ app/controllers/ems_infra_controller.rb
@@ -1,7 +1,7 @@
 class EmsInfraController < ApplicationController
   include Mixins::GenericListMixin
   include Mixins::GenericShowMixin
-  include EmsCommon # common methods for EmsInfra/Cloud controllers
+  include Mixins::EmsCommon # common methods for EmsInfra/Cloud controllers
   include Mixins::EmsCommonAngular
   include Mixins::DashboardViewMixin
