+++ app/controllers/ems_network_controller.rb
@@ -1,7 +1,7 @@
 class EmsNetworkController < ApplicationController
   include Mixins::GenericListMixin
   include Mixins::GenericShowMixin
-  include EmsCommon
+  include Mixins::EmsCommon
   include Mixins::EmsCommonAngular
   include Mixins::GenericSessionMixin
